concurrent: Abstract similarities in the futures classes.
